[backend][payment] - 결제 복구 서비스 기능 버그 픽스 및 테스트 케이스 작성

+++ back-end/src/main/java/com/example/backend/domain/PendingPaymentOrder.java
@@ -9,11 +9,11 @@ public class PendingPaymentOrder {
     private final Long paymentOrderId;
     private final PaymentStatus paymentStatus;
     private final Long amount;
-    private final int failedCount;
-    private final int threshold;
+    private final Byte failedCount;
+    private final Byte threshold;
 
     @Builder
-    public PendingPaymentOrder(Long paymentOrderId, PaymentStatus paymentStatus, Long amount, int failedCount, int threshold) {
+    public PendingPaymentOrder(Long paymentOrderId, PaymentStatus paymentStatus, Long amount, Byte failedCount, Byte threshold) {
         this.paymentOrderId = paymentOrderId;
         this.paymentStatus = paymentStatus;
         this.amount = amount;
